backend/requirements.txt
fastapi
uvicorn
pdfplumber
python-docx
spacy
scikit-learn
sentence-transformers
numpy
pandas